Disambiguate dwl-patches' license - closes issue #530
During the hosting transition from github to codeberg, dwl and
dwl-patches were split up for logistical reasons. In doing so,
however, the license files remained with dwl and were not replicated
into the new dwl-patches repository. This left new patches and
contributions to patches in an ambiguous license state.

Codeberg dwl-patches issue #530 requested that contributors to
dwl-patches explicitly consent or dissent to the dwl license
structure. All but seven contributors did so. Those seven
contributors' patches are here formally removed as dwl-patches adopts
the dwl license structure.

For clarity this license structure is as follows:

GPL-3+ license with a provision for the MIT license covering any
patches that affect util.c and util.h and a provision for the sway
license covering any patches that interact with client_is_float_type()
and the layer shell code which were drawn out of the sway code base.